Memoize warnings so modal doesn't close
